+++ client/src/App.js
@@ -26,7 +26,7 @@ import SingleFriend from './pages/SingleFriend';
 import Donate from './pages/Donate';
 import Success from './pages/Success';
 import SearchBar from './components/SearchBar';
-import SignUp from './pages/signUp';
+import SignupForm from './pages/SignupForm';
 import ProfileUpdate from './pages/ProfileUpdate';
 // import UserProfile from './components/UserProfile';
 import SideBar from './components/Sidebar';
@@ -100,7 +100,7 @@ function App() {
                   />
                   <Route
                     path="/signup"
-                    element={<SignUp />}
+                    element={<SignupForm />}
                   />
                   <Route
                     path='/logout'
